+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - [2020-04-08]
+### Added
+- Added EligibilityBenefit#medicare?
+- Added EligibilityData#medicare?
+
+### Changed
+- EligibilityBenefit(s)#individual returns true for medicare in benefit and queries that look for 'IND'
+- EligibilityBenefit(s)#in\_network returns true for medicare in benefit and queries that look for 'Y'
+
 ## [0.8.0] - [2020-04-04]
 ### Changed
-- Attemp to serialize all properties with 'date' in the name to ChangeHealth date format
+- Attempt to serialize all properties with 'date' in the name to ChangeHealth date format
 
 ## [0.7.0] - [2020-04-03]
 ### Changed